+++ docs/CHANGELOG.rst
@@ -16,6 +16,7 @@ Added
 - Add version to annotation field
 - Add a resolwe method for fetching the server module versions
 - Add support for ``Python 3.13``
+- Add serializer support to simplify custom serialization of resources
 
 Changed
 -------